requirements.txt
flask
flask-sqlalchemy
mysql-connector-python
python-dotenv
flask-swagger-ui